Refactor compare functions registration (#306)
